Flytte uthenting av organisasjoner brukeren har til fra Main (auth) til endepunktet request skal til
